Detect E3SM-Unified version and dev labels in deployment
To share infrastructure, the functions for detecting the
E3SM-Unified version and building a list of dev labels has been
moved to `shared.py`, which is now symlinked into the recipe
directory and used in `build_packages.py`
